Makes prp_check_img_exists() always return false.
The plugin does not have access to the WordPress SVN server, so any methods it calls to find the file will fail.

Possible workarounds are to use the PHP svn library or the PEAR VersionControl_SVN package, but there is no guarantee either of these will be installed on the user's server.