Fixes display of 'exclude=meta', 'exclude=head', include=head'.
Does not fix 'include=meta' or no parameters proveded, when the entire readme should be displayed.

Undoes much of commit 9ae4958, which ended up having the wrong logic.

Refactors the code such that the massive if-else block that does the head meta data is moved into a function in functions.php."