Moves most of the functions in functions.php to Generate_Output.
There's no point in most of them being separate, especially those that use a lot of data private to Generate_Output.

The functions that are left either output data, mainly for debug purposes, or have WordPress hooks related to them. There's also the complicated and weird affair that is the value returned by prp_toggle_global_shortcodes() to Generate_Output::toggle_global_shortcodes().

To be investigated.