Ensures prp_log() is not used for error messages.
The $error argument of prp_log() has been deprecated, but not yet removed because there's another (optional) argument following it. The default value of the deprecated argument, $deprecated_param, is false, as it was for $error. If true is passed, an exception is thrown.

The formatting of displayed errors is standardised.